+++ tests/weak.rs
@@ -1,4 +1,4 @@
-use simple_actors::{Actor, Context, Handler, Message};
+use simple_actors::{Actor, Context, Handle, Handler, Message};
 use test_log::test;
 
 #[derive(Default)]
@@ -6,6 +6,11 @@ struct MyActor {}
 
 impl Actor for MyActor {
     type Error = String;
+    type Args = ();
+
+    async fn create(_: Context, _: Handle<Self>, _: Self::Args) -> Result<Self, Self::Error> {
+        Ok(Self::default())
+    }
 }
 
 struct MyMessage {}